moved producer related kafka code under the `producer` subpackage
